feat(e2e): bump test utils and simplify beforeEach
