+++ workspaces/bulk-import/e2e-tests/package.json
@@ -24,7 +24,7 @@
   "devDependencies": {
     "@eslint/js": "^9.39.2",
     "@playwright/test": "^1.56.1",
-    "@red-hat-developer-hub/e2e-test-utils": "1.1.32",
+    "@red-hat-developer-hub/e2e-test-utils": "1.1.43",
     "@types/node": "^24.10.1",
     "dotenv": "^16.4.7",
     "eslint": "^9.39.2",
