+++ workspaces/bulk-import/e2e-tests/tests/support/utils.ts
@@ -1,42 +1,17 @@
-import { LoginHelper } from "@red-hat-developer-hub/e2e-test-utils/helpers";
+import { LoginHelper, UIhelper } from "@red-hat-developer-hub/e2e-test-utils/helpers";
 import { expect, Page } from "@red-hat-developer-hub/e2e-test-utils/test";
 
-export async function ensureGithubUserSession(
+/** Bulk Import "Login Required" — only await reauth when Log in opens a popup (utils listener has no timeout). */
+async function dismissBulkImportLoginDialogIfPresent(
   page: Page,
   loginHelper: LoginHelper,
-): Promise<void> {
-  await page.goto("/");
-  const appReady = await page
-    .locator("nav a")
-    .first()
-    .isVisible({ timeout: 15_000 })
-    .catch(() => false);
-  if (appReady) {
-    return; // cookies/session already valid for this page
-  }
-  await loginHelper.loginAsGithubUser();
-}
-
-type GithubLoginHelper = {
-  checkAndReauthorizeGithubApp: () => Promise<void>;
-};
-
-/**
- * Bulk Import shows "Login Required" after the page content paints. Wait for the
- * dialog (do not use a one-shot isVisible right after the page marker).
- */
-export async function dismissBulkImportLoginDialogIfPresent(
-  page: Page,
-  loginHelper: GithubLoginHelper,
   waitForDialogMs = 8_000,
 ): Promise<void> {
   const loginDialog = page.getByRole("dialog", { name: "Login Required" });
-
   const appeared = await loginDialog
     .waitFor({ state: "visible", timeout: waitForDialogMs })
     .then(() => true)
     .catch(() => false);
-
   if (!appeared) {
     return;
   }
@@ -44,11 +19,52 @@ export async function dismissBulkImportLoginDialogIfPresent(
   const logInButton = loginDialog.getByRole("button", { name: "Log in" });
   await expect(logInButton).toBeVisible({ timeout: 10_000 });
 
-  const authorize = loginHelper.checkAndReauthorizeGithubApp();
-  await logInButton.click();
-  await authorize;
+  const reauthorize = loginHelper.checkAndReauthorizeGithubApp();
+  const popup = await Promise.all([
+    page.waitForEvent("popup", { timeout: 15_000 }),
+    logInButton.click(),
+  ])
+    .then(([p]) => p)
+    .catch(() => null);
+
+  if (popup) {
+    await reauthorize;
+  }
 
   await expect(loginDialog).toBeHidden({ timeout: 60_000 });
+}
+
+/**
+ * GitHub RHDH session (when needed) + navigate to Bulk import and clear SCM login dialog.
+ */
+export async function prepareBulkImportPage(
+  page: Page,
+  loginHelper: LoginHelper,
+  uiHelper: UIhelper,
+): Promise<void> {
+  await page.goto("/");
+
+  if (
+    await page
+      .getByRole("button", { name: "Sign In" })
+      .isVisible({ timeout: 15_000 })
+      .catch(() => false)
+  ) {
+    await loginHelper.loginAsGithubUser();
+  }
+
+  await uiHelper.openSidebar("Bulk import");
+
+  const bulkImportReady = page.getByText("Source control tool", {
+    exact: true,
+  });
+  await expect(bulkImportReady).toBeVisible({ timeout: 20_000 });
+  await dismissBulkImportLoginDialogIfPresent(page, loginHelper);
+  await expect(
+    page.getByRole("dialog", { name: "Login Required" }),
+  ).toBeHidden({ timeout: 5_000 });
+  await expect(bulkImportReady).toBeVisible();
+  await uiHelper.verifyHeading("Bulk import");
 }
 
 const GITLAB_LOGIN_REJECTED_EMPTY_STATE = "Log in to view projects";
